Feat: Add assets api
